[WiP] all tests send auth headers and main routes require JWT. https://github.com/ideaq/time/issues/54
