+++ README.md
@@ -10,7 +10,7 @@
 
 - - -
 
-## Why?
+## Why? [![Start with Why](https://img.shields.io/badge/start%20with-why%3F-brightgreen.svg?style=flat)](http://www.ted.com/talks/simon_sinek_how_great_leaders_inspire_action)
 
 > “*The **price** of anything is the **amount** of **life** you **exchange for it**.*”
 ~ Henry David Thoreau
@@ -29,9 +29,11 @@ http://youtu.be/tvos4nORf_Y
 
 ## What?
 
+
+
 A simple way to track activities over time.
 
-## How?
+## How? [![contributions welcome](https://img.shields.io/badge/contributions-welcome-brightgreen.svg?style=flat)](https://github.com/ideaq/time/issues)
 
 > Help us build it! https://github.com/ideaq/time/issues
