+++ handlers/auth_jwt_validate.js
@@ -1,9 +1,9 @@
 // bring your own validation function
 var validateFunc = function (decoded, callback) {
 
-    console.log("- - - - JWT - - - - ");
-    console.log(decoded);
-    console.log("- - - - - - - - - - ");
+    // console.log("- - - - JWT - - - - ");
+    // console.log(decoded);
+    // console.log("- - - - - - - - - - ");
 
     // do your checks to see if the person is valid
     // if (!people[decoded.id]) {
